src/third_party/blink/web_tests/external/wpt/mathml/relations/css-styling/default-properties-on-the-math-root-expected.txt

This is a testharness.js-based test.
FAIL Value of direction on the <math> root assert_equals: when specified from the UA sheet expected "ltr" but got "rtl"
PASS Value of writing-mode on the <math> root
FAIL Value of text-indent on the <math> root assert_equals: when specified from the UA sheet expected "0px" but got "8px"
FAIL Value of letter-spacing on the <math> root assert_equals: when specified from the UA sheet expected "normal" but got "8px"
FAIL Value of line-height on the <math> root assert_equals: when specified from the UA sheet expected "normal" but got "8px"
FAIL Value of word-spacing on the <math> root assert_equals: when specified from the UA sheet expected "0px" but got "8px"
FAIL Value of font-style on the <math> root assert_equals: when specified from the UA sheet expected "normal" but got "italic"
FAIL Value of font-weight on the <math> root assert_equals: when specified from the UA sheet expected "400" but got "700"
Harness: the test ran to completion.
